--- a/benchmarks/build.gradle.kts
+++ b/benchmarks/build.gradle.kts
@@ -11,6 +11,9 @@ plugins {
 dependencies {
     // Make jmhCompileClasspath resolvable
     jmhImplementation(project(":postgresql"))
+    jmhRuntimeOnly("com.ongres.scram:scram-client:3.1") {
+        because("We refer non-shaded :postgresql, so we add explicit scram-client dependency")
+    }
     jmhImplementation(testFixtures(project(":postgresql")))
     jmhImplementation("org.roaringbitmap:RoaringBitmap:1.3.0")
     jmhImplementation("it.unimi.dsi:fastutil:8.5.15")