@vlsi 2025-05-30
Any thoughts?
I think we could merge this and release yet another version to get the perf issue fixed.

The core changes here are like `String->StringBuilder`. It would be great to add `try-with-resources`, however I don't like to mix many changes into a single PR.